requirements.txt
beautifulsoup4==4.14.2
openai==2.31.0
python-dotenv==1.2.2
requests==2.32.5
streamlit==1.56.0
